日付・時間

+++ models/product.py
@@ -1,9 +1,9 @@
-from peewee import Model, CharField, DecimalField
+from peewee import Model, CharField, IntegerField
 from .db import db
 
 class Product(Model):
     name = CharField()
-    price = DecimalField()
+    price = IntegerField()
 
     class Meta:
         database = db
